tools/dav-script/script/requirements.txt
mysql-connector-python
pytz
datetime
jira
openpyxl
pyyaml